minor code cleanup and major TODO list commenting

+++ include/project/chassis.h
@@ -11,13 +11,13 @@ public:
     void user_control();
     void set(float power, float turn, float strafe);
 
-    double power_mult_calc();
+    float power_mult_calc();
 
     int vision_align();
     std::unique_ptr<pros::Task> align_task;
 
-    double turn_radius;
-    double wheel_circumference;
+    float turn_radius;
+    float wheel_circumference;
     float motor_speed;
 
 protected:
